requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
paramiko==3.3.1
psycopg2-binary==2.9.7
redis==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-environ==0.11.2
celery==5.3.4
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
daphne==4.0.0
websockets==11.0.3
cryptography==41.0.7
django-widget-tweaks==1.5.0
pillow==10.1.0